Rename simple_translate.py to architecture.py
Point imports and Modal packaging at the new module; update README and app link.

Made-with: Cursor
