+++ architecture.py
